Improved /tps accuracy

+++ src/common/me/nallar/tickthreading/minecraft/commands/TPSCommand.java
@@ -27,7 +27,7 @@ public class TPSCommand extends Command {
 		for (TickManager tickManager : TickThreading.instance.getManagers()) {
 			tpsReport.append(tickManager.getBasicStats());
 		}
-		int usedTime = MinecraftServer.getTickTime();
+		float usedTime = MinecraftServer.getTickTime();
 		tpsReport.append("\nUsed time per tick: ").append(usedTime).append("ms")
 				.append("\nOverall TPS: ").append(Math.min(20, 1000 / (usedTime == 0 ? 1 : usedTime)))
 				.append("\nOverall load: ").append(usedTime * 2).append('%');
